+++ src/main/java/com/techfork/post/infrastructure/batch/PostSummaryReader.java
@@ -1,4 +1,4 @@
-package com.techfork.post.application.batch;
+package com.techfork.post.infrastructure.batch;
 
 import com.techfork.post.domain.Post;
 import com.techfork.post.infrastructure.PostRepository;
