+++ src/test/java/com/techfork/domain/source/config/RssCrawlingJobIntegrationTest.java
@@ -1,11 +1,11 @@
 package com.techfork.domain.source.config;
 
 import com.techfork.post.application.batch.PostEmbeddingProcessor;
-import com.techfork.post.application.batch.PostEmbeddingReader;
-import com.techfork.post.application.batch.PostEmbeddingWriter;
+import com.techfork.post.infrastructure.batch.PostEmbeddingReader;
+import com.techfork.post.infrastructure.batch.PostEmbeddingWriter;
 import com.techfork.post.application.batch.PostSummaryProcessor;
-import com.techfork.post.application.batch.PostSummaryReader;
-import com.techfork.post.application.batch.PostSummaryWriter;
+import com.techfork.post.infrastructure.batch.PostSummaryReader;
+import com.techfork.post.infrastructure.batch.PostSummaryWriter;
 import com.techfork.post.domain.projection.ContentChunk;
 import com.techfork.post.domain.projection.PostDocument;
 import com.techfork.post.domain.Post;
