+++ src/test/java/com/techfork/post/infrastructure/batch/PostEmbeddingReaderDataJpaTest.java
@@ -1,12 +1,11 @@
-package com.techfork.post.application.batch;
+package com.techfork.post.infrastructure.batch;
 
 import com.techfork.post.domain.Post;
+import com.techfork.post.infrastructure.PostRepository;
 import com.techfork.domain.source.dto.RssFeedItem;
 import com.techfork.domain.source.entity.TechBlog;
 import com.techfork.domain.source.repository.TechBlogRepository;
-import com.techfork.post.infrastructure.PostRepository;
 import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceUnitUtil;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -14,6 +13,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.util.ReflectionTestUtils;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -23,7 +23,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @DataJpaTest
 @ActiveProfiles("test")
-class PostSummaryReaderDataJpaTest {
+class PostEmbeddingReaderDataJpaTest {
 
     @Autowired
     private PostRepository postRepository;
@@ -53,20 +53,23 @@ class PostSummaryReaderDataJpaTest {
     class Read {
 
         @Test
-        @DisplayName("summary가 null이거나 빈 문자열인 게시글만 읽는다")
-        void readsOnlyPostsWithNullOrEmptySummary() throws Exception {
-            Post nullSummaryPost = savePost("null-summary", null, null, List.of("AI"));
-            Post emptySummaryPost = savePost("empty-summary", "", "", List.of("Batch"));
-            savePost("completed-summary", "완료 요약", "완료 짧은 요약", List.of("Done"));
+        @DisplayName("요약이 공백이 아니고 embeddedAt이 null인 게시글만 읽는다")
+        void readsOnlyPostsReadyForEmbedding() throws Exception {
+            Post readyPost1 = savePost("ready-1", "요약 완료 1", "짧은 요약 1", null);
+            Post readyPost2 = savePost("ready-2", "요약 완료 2", "짧은 요약 2", null);
+            savePost("null-summary", null, null, null);
+            savePost("empty-summary", "", "", null);
+            savePost("blank-summary", "   ", "   ", null);
+            savePost("already-embedded", "이미 임베딩됨", "짧은 요약", LocalDateTime.of(2026, 5, 11, 9, 0));
 
             entityManager.clear();
 
-            PostSummaryReader postSummaryReader = new PostSummaryReader(postRepository);
+            PostEmbeddingReader postEmbeddingReader = new PostEmbeddingReader(postRepository);
             List<Post> readPosts = new ArrayList<>();
 
-            Post firstRead = postSummaryReader.read();
-            Post secondRead = postSummaryReader.read();
-            Post thirdRead = postSummaryReader.read();
+            Post firstRead = postEmbeddingReader.read();
+            Post secondRead = postEmbeddingReader.read();
+            Post thirdRead = postEmbeddingReader.read();
 
             readPosts.add(firstRead);
             readPosts.add(secondRead);
@@ -74,44 +77,34 @@ class PostSummaryReaderDataJpaTest {
             assertThat(thirdRead).isNull();
             assertThat(readPosts)
                     .extracting(Post::getId)
-                    .containsExactlyInAnyOrder(nullSummaryPost.getId(), emptySummaryPost.getId());
-
-            PersistenceUnitUtil persistenceUnitUtil = entityManager.getEntityManagerFactory().getPersistenceUnitUtil();
+                    .containsExactlyInAnyOrder(readyPost1.getId(), readyPost2.getId());
             assertThat(readPosts)
-                    .allSatisfy(post -> assertThat(persistenceUnitUtil.isLoaded(post, "keywords")).isTrue());
-            assertThat(readPosts.stream()
-                    .filter(post -> post.getId().equals(nullSummaryPost.getId()))
-                    .findFirst()
-                    .orElseThrow()
-                    .getKeywords())
-                    .extracting(keyword -> keyword.getKeyword())
-                    .containsExactly("AI");
-            assertThat(readPosts.stream()
-                    .filter(post -> post.getId().equals(emptySummaryPost.getId()))
-                    .findFirst()
-                    .orElseThrow()
-                    .getKeywords())
-                    .extracting(keyword -> keyword.getKeyword())
-                    .containsExactly("Batch");
+                    .allSatisfy(post -> {
+                        assertThat(post.getSummary()).isNotBlank();
+                        assertThat(post.getEmbeddedAt()).isNull();
+                    });
         }
 
         @Test
-        @DisplayName("summary가 있는 게시글만 있으면 null을 반환한다")
-        void returnsNullWhenNoPostsMatchSummaryCondition() throws Exception {
-            savePost("completed-summary", "완료 요약", "완료 짧은 요약", List.of("Done"));
+        @DisplayName("조건을 만족하는 게시글이 없으면 null을 반환한다")
+        void returnsNullWhenNoPostsAreReadyForEmbedding() throws Exception {
+            savePost("null-summary", null, null, null);
+            savePost("empty-summary", "", "", null);
+            savePost("blank-summary", "   ", "   ", null);
+            savePost("already-embedded", "이미 임베딩됨", "짧은 요약", LocalDateTime.of(2026, 5, 11, 9, 0));
 
             entityManager.clear();
 
-            PostSummaryReader postSummaryReader = new PostSummaryReader(postRepository);
+            PostEmbeddingReader postEmbeddingReader = new PostEmbeddingReader(postRepository);
 
-            assertThat(postSummaryReader.read()).isNull();
+            assertThat(postEmbeddingReader.read()).isNull();
         }
     }
 
-    private Post savePost(String suffix, String summary, String shortSummary, List<String> keywords) {
+    private Post savePost(String suffix, String summary, String shortSummary, LocalDateTime embeddedAt) {
         Post post = Post.create(
                 RssFeedItem.builder()
-                        .title("요약 대상 글 " + suffix)
+                        .title("임베딩 대상 글 " + suffix)
                         .url("https://posts.example.com/" + suffix)
                         .logoUrl("https://cdn.example.com/logo-" + suffix + ".png")
                         .thumbnailUrl("https://cdn.example.com/thumb-" + suffix + ".png")
@@ -124,7 +117,7 @@ class PostSummaryReaderDataJpaTest {
                 techBlog
         );
         post.updateSummaries(summary, shortSummary);
-        post.replaceKeywords(keywords);
+        ReflectionTestUtils.setField(post, "embeddedAt", embeddedAt);
         return postRepository.saveAndFlush(post);
     }
 }
