+++ src/test/java/com/techfork/post/infrastructure/batch/PostEmbeddingWriterTest.java
@@ -1,4 +1,4 @@
-package com.techfork.post.application.batch;
+package com.techfork.post.infrastructure.batch;
 
 import co.elastic.clients.elasticsearch.ElasticsearchClient;
 import co.elastic.clients.elasticsearch._types.ErrorCause;
