+++ src/test/java/com/techfork/post/infrastructure/batch/PostSummaryReaderDataJpaTest.java
@@ -1,11 +1,12 @@
-package com.techfork.post.application.batch;
+package com.techfork.post.infrastructure.batch;
 
 import com.techfork.post.domain.Post;
-import com.techfork.post.infrastructure.PostRepository;
 import com.techfork.domain.source.dto.RssFeedItem;
 import com.techfork.domain.source.entity.TechBlog;
 import com.techfork.domain.source.repository.TechBlogRepository;
+import com.techfork.post.infrastructure.PostRepository;
 import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceUnitUtil;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -13,7 +14,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.util.ReflectionTestUtils;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -23,7 +23,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @DataJpaTest
 @ActiveProfiles("test")
-class PostEmbeddingReaderDataJpaTest {
+class PostSummaryReaderDataJpaTest {
 
     @Autowired
     private PostRepository postRepository;
@@ -53,23 +53,20 @@ class PostEmbeddingReaderDataJpaTest {
     class Read {
 
         @Test
-        @DisplayName("요약이 공백이 아니고 embeddedAt이 null인 게시글만 읽는다")
-        void readsOnlyPostsReadyForEmbedding() throws Exception {
-            Post readyPost1 = savePost("ready-1", "요약 완료 1", "짧은 요약 1", null);
-            Post readyPost2 = savePost("ready-2", "요약 완료 2", "짧은 요약 2", null);
-            savePost("null-summary", null, null, null);
-            savePost("empty-summary", "", "", null);
-            savePost("blank-summary", "   ", "   ", null);
-            savePost("already-embedded", "이미 임베딩됨", "짧은 요약", LocalDateTime.of(2026, 5, 11, 9, 0));
+        @DisplayName("summary가 null이거나 빈 문자열인 게시글만 읽는다")
+        void readsOnlyPostsWithNullOrEmptySummary() throws Exception {
+            Post nullSummaryPost = savePost("null-summary", null, null, List.of("AI"));
+            Post emptySummaryPost = savePost("empty-summary", "", "", List.of("Batch"));
+            savePost("completed-summary", "완료 요약", "완료 짧은 요약", List.of("Done"));
 
             entityManager.clear();
 
-            PostEmbeddingReader postEmbeddingReader = new PostEmbeddingReader(postRepository);
+            PostSummaryReader postSummaryReader = new PostSummaryReader(postRepository);
             List<Post> readPosts = new ArrayList<>();
 
-            Post firstRead = postEmbeddingReader.read();
-            Post secondRead = postEmbeddingReader.read();
-            Post thirdRead = postEmbeddingReader.read();
+            Post firstRead = postSummaryReader.read();
+            Post secondRead = postSummaryReader.read();
+            Post thirdRead = postSummaryReader.read();
 
             readPosts.add(firstRead);
             readPosts.add(secondRead);
@@ -77,34 +74,44 @@ class PostEmbeddingReaderDataJpaTest {
             assertThat(thirdRead).isNull();
             assertThat(readPosts)
                     .extracting(Post::getId)
-                    .containsExactlyInAnyOrder(readyPost1.getId(), readyPost2.getId());
+                    .containsExactlyInAnyOrder(nullSummaryPost.getId(), emptySummaryPost.getId());
+
+            PersistenceUnitUtil persistenceUnitUtil = entityManager.getEntityManagerFactory().getPersistenceUnitUtil();
             assertThat(readPosts)
-                    .allSatisfy(post -> {
-                        assertThat(post.getSummary()).isNotBlank();
-                        assertThat(post.getEmbeddedAt()).isNull();
-                    });
+                    .allSatisfy(post -> assertThat(persistenceUnitUtil.isLoaded(post, "keywords")).isTrue());
+            assertThat(readPosts.stream()
+                    .filter(post -> post.getId().equals(nullSummaryPost.getId()))
+                    .findFirst()
+                    .orElseThrow()
+                    .getKeywords())
+                    .extracting(keyword -> keyword.getKeyword())
+                    .containsExactly("AI");
+            assertThat(readPosts.stream()
+                    .filter(post -> post.getId().equals(emptySummaryPost.getId()))
+                    .findFirst()
+                    .orElseThrow()
+                    .getKeywords())
+                    .extracting(keyword -> keyword.getKeyword())
+                    .containsExactly("Batch");
         }
 
         @Test
-        @DisplayName("조건을 만족하는 게시글이 없으면 null을 반환한다")
-        void returnsNullWhenNoPostsAreReadyForEmbedding() throws Exception {
-            savePost("null-summary", null, null, null);
-            savePost("empty-summary", "", "", null);
-            savePost("blank-summary", "   ", "   ", null);
-            savePost("already-embedded", "이미 임베딩됨", "짧은 요약", LocalDateTime.of(2026, 5, 11, 9, 0));
+        @DisplayName("summary가 있는 게시글만 있으면 null을 반환한다")
+        void returnsNullWhenNoPostsMatchSummaryCondition() throws Exception {
+            savePost("completed-summary", "완료 요약", "완료 짧은 요약", List.of("Done"));
 
             entityManager.clear();
 
-            PostEmbeddingReader postEmbeddingReader = new PostEmbeddingReader(postRepository);
+            PostSummaryReader postSummaryReader = new PostSummaryReader(postRepository);
 
-            assertThat(postEmbeddingReader.read()).isNull();
+            assertThat(postSummaryReader.read()).isNull();
         }
     }
 
-    private Post savePost(String suffix, String summary, String shortSummary, LocalDateTime embeddedAt) {
+    private Post savePost(String suffix, String summary, String shortSummary, List<String> keywords) {
         Post post = Post.create(
                 RssFeedItem.builder()
-                        .title("임베딩 대상 글 " + suffix)
+                        .title("요약 대상 글 " + suffix)
                         .url("https://posts.example.com/" + suffix)
                         .logoUrl("https://cdn.example.com/logo-" + suffix + ".png")
                         .thumbnailUrl("https://cdn.example.com/thumb-" + suffix + ".png")
@@ -117,7 +124,7 @@ class PostEmbeddingReaderDataJpaTest {
                 techBlog
         );
         post.updateSummaries(summary, shortSummary);
-        ReflectionTestUtils.setField(post, "embeddedAt", embeddedAt);
+        post.replaceKeywords(keywords);
         return postRepository.saveAndFlush(post);
     }
 }
